The need for these optimizations was discovered through the use of Devel::NYTProf
while running a 10k-line unfiltered GL search report. These optimizations together
reduce the runtime of the report from 425secs to 275secs.

Please do note that I consider 275secs still disappointing, but a lot less disappointing
than 425secs.

Optimizations:

 * Memoize output formatters in PGDate and PGNumber
 * Restructure LedgerSMB::Template::preprocess()
 * Checking for defined-ness (//) is much cheaper than
   checking for truth (||) with PGNumber
 * Checking it's safe to call `can` or `isa` is much
   cheaper than catching failure using eval{}